+++ core/libraries/bancha/view.php
@@ -151,8 +151,11 @@ Class View
 
 		//Theme hooks - added in Bancha 2.0
 		$hooks = THEMESPATH . $this->theme . DIRECTORY_SEPARATOR . 'hooks.php';
-		if (file_exists($hooks)) {
+		if (THEME_HOOKS && file_exists($hooks)) {
 			include_once($hooks);
+			if (function_exists('hook_onconstruct')) {
+				hook_onconstruct();
+			}
 		}