+++ themes/teabag/hooks.php
@@ -1,5 +1,35 @@
 <?php
+/**
+ * Theme Hooks
+ *
+ * Please declare only functions that you actually use.
+ *
+ * @package		Bancha
+ * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
+ * @copyright	Copyright (c) 2011-2012, Squallstar
+ * @license		GNU/GPL (General Public License)
+ * @link		http://squallstar.it
+ *
+ */
 
+/**
+ * Called as soon as the view class is loaded
+ */
+function hook_onconstruct() {
+	//echo 'Hello world';
+}
+
+/**
+ * Called when the controller is destructed
+ */
+function hook_ondestruct() {
+	//echo 'Hello world';
+}
+
+/**
+ * Administration logo override
+ * @return string
+ */
 function hook_admin_logo() {
-	return 'Hello';
+	//return 'Hello world';
 }