+++ core/documentation/source/basic/changelog.rst
@@ -4,6 +4,7 @@ ChangeLog
 
 **v 2.0** (Not planned - development version)
 
+- Theme hooks: now themes can register many functions as hooks that will be called like Wordpress do
 - Added an install autoredirect on the settings model
 
 **v 1.2.2** (2012-06-14)